# Conflicts:
#	app/build.gradle

+++ app/src/main/java/com/app/didaktikapp/BBDD/Service/GrupoService.java
@@ -37,7 +37,7 @@ public class GrupoService {
         return grupoDao.getGrupos();
     }
 
-    public Grupo getGrupo(Integer id) {
+    public Grupo getGrupo(Long id) {
         return grupoDao.getGrupo(id);
     }
